% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{BH}
\alias{BH}
\title{BH}
\usage{
BH(SSB, B0, h)

BH(SSB, B0, h)
}
\arguments{
\item{SSB}{Spawning stock biomass}

\item{B0}{equilibrium SSB}

\item{h}{steepness parameter as defined by Doonan and Mace 1981, represents}
}
\value{
Number of recruits acording to the Beverton holt relationship

Number of recruits acording to the Beverton holt relationship
}
\description{
Beverton holt stock recruit relationship parameterised with steepness

Beverton holt stock recruit relationship parameterised with steepness
}
