% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{bound_unit}
\alias{bound_unit}
\title{bound_unit constrains Y from -inf -> inf to be between -1 -> 1}
\usage{
bound_unit(Y)
}
\arguments{
\item{Y}{scalar range \link{-inf, inf}}
}
\value{
X to be between \link{-1,1}
}
\description{
bound_unit constrains Y from -inf -> inf to be between -1 -> 1
}
