% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{calculate_initial_numbers_at_age_age_based_movement}
\alias{calculate_initial_numbers_at_age_age_based_movement}
\title{calculate_initial_numbers_at_age_age_based_movement}
\usage{
calculate_initial_numbers_at_age_age_based_movement(
  n_regions,
  n_ages,
  R0,
  old_movement_matrix,
  young_movement_matrix,
  age_based_movement_ogive,
  natural_mortality
)
}
\arguments{
\item{n_regions}{integer number of regions}

\item{n_ages}{integer number of ages}

\item{R0}{vector of R0 parameters for each region}

\item{old_movement_matrix}{movement matrix for older fish}

\item{young_movement_matrix}{movement matrix for young fish}

\item{age_based_movement_ogive}{age_based movement for young fish}

\item{natural_mortality}{vector of natural mortality rates for each age}
}
\value{
matrix of numbers at age with regions being rows and ages being the cols
}
\description{
calculate_initial_numbers_at_age_age_based_movement
}
