% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{calculate_simulated_residuals}
\alias{calculate_simulated_residuals}
\title{calculate_simulated_residuals}
\usage{
calculate_simulated_residuals(sim_ob, type = "abundance")
}
\arguments{
\item{sim_ob}{a data frame which should be one of the elements that are created from \code{simulate_observations}}

\item{type}{string defining type
\itemize{
\item \code{abundance}: abundance
\item \code{AF}: Age frequency
\item \code{LF}: Length frequency
\item \code{tag-recovery}: tag recovery observations
}}
}
\value{
data.frame of scaled simulated residuals for each release and recovery event
}
\description{
calculate_simulated_residuals
}
\details{
take simulated observations created from \code{simulate_observations} and creating simulated scaled residuals using the DHARMa package \insertCite{dharma}{SpatialSablefishAssessment}
}
\references{
\insertAllCited{}
}
