% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInputValidation.R
\name{check_dim}
\alias{check_dim}
\title{check_dim checks the dimension of an array are consistent with expected values}
\usage{
check_dim(array, expected_dimensions)
}
\arguments{
\item{array}{a matrix or array that we are checking}

\item{expected_dimensions}{a vector with the expected number of dimensions that the array should have}
}
\value{
a list
}
\description{
check_dim checks the dimension of an array are consistent with expected values
}
