% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{check_gradients}
\alias{check_gradients}
\title{check_gradients}
\usage{
check_gradients(obj)
}
\arguments{
\item{obj}{A TMB list that has been built by MakeAdFun}
}
\value{
bool and print a message. True is gradients all non zero false suggests there are parameters with zero gradients which is a problem
}
\description{
check_gradients
}
\details{
checks a TMB object for fixed effect parameters that have 0 gradients, suggesting they don't contribute the log likelihood. And you should
look into these parameters.
}
