% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInputValidation.R
\name{check_length}
\alias{check_length}
\title{check_length checks the length of vector with expected values}
\usage{
check_length(vector_to_check, n_elements)
}
\arguments{
\item{vector_to_check}{a matrix or array that we are checking}

\item{n_elements}{the expected length of vectors}
}
\value{
a list
}
\description{
check_length checks the length of vector with expected values
}
