% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{constant_sel}
\alias{constant_sel}
\title{constant_sel constant selectivity}
\usage{
constant_sel(C, bins)
}
\arguments{
\item{C}{values}

\item{bins}{could be age or length}
}
\value{
constant selectivty
}
\description{
constant_sel constant selectivity
}
