% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{convert_simdata_integers}
\alias{convert_simdata_integers}
\title{convert_simdata_integers}
\usage{
convert_simdata_integers(sim_data, OM_data)
}
\arguments{
\item{sim_data}{an array with 3 dimensions}

\item{OM_data}{data that was passed to \code{MakeADFun}}
}
\value{
sim_data list with integers converted
}
\description{
convert_simdata_integers
}
\details{
this function should be used if you get the following error from \code{MakeADFun} \verb{Error in getParameterOrder(data, parameters, new.env(), DLL = DLL) : NOT A VECTOR!}
}
