% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{d_exp_sel}
\alias{d_exp_sel}
\title{d_exp_sel Double Exponential selectivity}
\usage{
d_exp_sel(bins, x_1, x_2, x_0, y_0, y_1, y_2)
}
\arguments{
\item{bins}{could be age or length}

\item{x_1}{reference bin for the left hand point}

\item{x_2}{reference bin for the right hand point}

\item{x_0}{reference bin for middle point}

\item{y_0}{selectivty at middle bin}

\item{y_1}{selectivity at left hand bin}

\item{y_2}{selectivity at right hand bin}
}
\description{
d_exp_sel Double Exponential selectivity
}
