% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{d_norm_sel}
\alias{d_norm_sel}
\title{d_norm_sel Double normal selectivity}
\usage{
d_norm_sel(bins, mu, sig_l, sig_r)
}
\arguments{
\item{bins}{could be age or length}

\item{mu}{mean value selectivity at one}

\item{sig_l}{sigma for the left hand curve}

\item{sig_r}{sigma for the right hand curve}
}
\value{
selectivity for each age
}
\description{
d_norm_sel Double normal selectivity
}
