% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{ddirichmult}
\alias{ddirichmult}
\title{ddirichmult
Return the PDF for the dirichlet-multinomial distribution based on \insertCite{thorson2017model}{SpatialSablefishAssessment}}
\usage{
ddirichmult(obs, beta, n, est, log_it = F)
}
\arguments{
\item{obs, }{vector of observed compositions assumes sum(obs) = 1}

\item{beta, }{the variance inflation coefficient (beta = n * theta)}

\item{n, }{the total number of samples in the available data (which is restricted to any non-negative real number),}

\item{est, }{vector of fiited compositions assumes sum(est) = 1}

\item{log_it}{return log pdf}
}
\value{
pdf
}
\description{
ddirichmult
Return the PDF for the dirichlet-multinomial distribution based on \insertCite{thorson2017model}{SpatialSablefishAssessment}
}
\details{
\insertCite{thorson2017model}{SpatialSablefishAssessment} deviates from the classic formulation by pulling out (x!) of the denominator from the right hand product
and moving it to the left denominator has been validated against extraDistr::ddirmnom
}
\references{
\insertAllCited{}
}
