% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dmultinom_upd}
\alias{dmultinom_upd}
\title{dmultinom_upd
an updated version of the multinomial so that we can test TMB's multinomial call}
\usage{
dmultinom_upd(x, prob, log = F)
}
\arguments{
\item{prob}{vector of fitted compositions assumes sum(prob) = 1}

\item{log}{return log of of the pdf}

\item{obs}{vector of observed values sum = N-effective}
}
\value{
PDF or log of the pdf
}
\description{
dmultinom_upd
an updated version of the multinomial so that we can test TMB's multinomial call
}
