% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{double_norm_ss3}
\alias{double_norm_ss3}
\title{double_norm_ss3 Double normal selectivity from SS3}
\usage{
double_norm_ss3(bin_vec, p1, p2, p3, p4, p5, p6)
}
\arguments{
\item{bin_vec}{could be age or length}

\item{p1}{peak parameter beginning size (or age) for the plateau (in cm or year).}

\item{p2}{width of plateau, as logistic between peak and maximum length (or age)}

\item{p3}{Ascending width: parameter value is ln(width).}

\item{p4}{Descending width: parameter value is ln(width).}

\item{p5}{y0 logistic transformed selectivity at first bin}

\item{p6}{y1 logistic transformed selectivity at last bin}
}
\description{
double_norm_ss3 Double normal selectivity from SS3
}
