% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{double_normal}
\alias{double_normal}
\title{double_normal a double normal selectivity}
\usage{
double_normal(x, x50, delta, amax = 1)
}
\arguments{
\item{x}{age or length to evaluate selectivity at}

\item{x50}{value of x where selectivity is equal to 0.5}

\item{delta}{delta parameter for the}
}
\value{
selectivity along x
}
\description{
double_normal a double normal selectivity
}
