% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{estimate_with_phases}
\alias{estimate_with_phases}
\title{Estimate a model with phases using phases}
\usage{
estimate_with_phases(data, parameters, map, phases, optimizer = "nlminb")
}
\arguments{
\item{data}{A list to be passed to TMB}

\item{parameters}{A list of parameters of the model}

\item{phases}{A list of the phases for the parameters of the model (same structure as your parameter list)}

\item{optimizer}{The optimizer to use. Default is nlminb (This is not currently active)}

\item{model_name}{A string describing the model name. Must be the name of your .cpp file}
}
\value{
A list of parameter estimates and their standard errors
}
\description{
This function runs TMB with ADMB-like phasing of parameter estimation.
Phasing within an optimizer
Function with normal inputs, passed via “...”, plus two additional arguments, “phase” and “optimizer”
Optimizer by default is nlminb
phase is a tagged list where missing elements are populated with a vector of 1s, and non-missing elements are integers, and where the optimizer loops through values of phase while progressively changing map to turn on parameters
}
\details{
see here https://github.com/kaskr/TMB_contrib_R/tree/master/TMBphase for information, was taken from TMBphase, but adjusted for our case which is fairly edge
}
\author{
Gavin Fay
}
