% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{evalit}
\alias{evalit}
\title{Utility function for executing a string in R}
\usage{
evalit(x)
}
\arguments{
\item{x}{string}
}
\value{
will exectute the string
}
\description{
Utility function for executing a string in R
}
\author{
Craig Marsh
}
