% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{every_nth}
\alias{every_nth}
\title{every_nth}
\usage{
every_nth(n)
}
\arguments{
\item{n}{every 'n' values will equal TRUE}
}
\value{
bool
}
\description{
return a vector of boolean that has every nth value = true
}
