% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{exp_sel}
\alias{exp_sel}
\title{exp_sel Exponential selectivity}
\usage{
exp_sel(bins, lambda)
}
\arguments{
\item{bins}{could be age or length}

\item{lambda}{rate decrease}
}
\value{
selectivity for each age.
}
\description{
exp_sel Exponential selectivity
}
