% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{extend_2darray}
\alias{extend_2darray}
\title{extend_2darray}
\usage{
extend_2darray(array_2d, n, colwise = TRUE)
}
\arguments{
\item{array_2d}{an array with 2 dimensions or matrix}

\item{n}{number of times to concatenate the last element in the 2 dimension}

\item{colwise}{boolean whether to append columns (TRUE) or rows (FALSE)}
}
\value{
a two dimensional array that has had the either the first or second dimension concatenated
}
\description{
extend_2darray
}
