% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{extend_3darray_last_dim}
\alias{extend_3darray_last_dim}
\title{extend_3darray_last_dim}
\usage{
extend_3darray_last_dim(array_3d, n)
}
\arguments{
\item{array_3d}{an array with 3 dimensions}

\item{n}{number of times to concatenate the last element in the 3 dimension}
}
\value{
a three dimensional array that has had the 3dimension concatenated
}
\description{
extend_3darray_last_dim
}
