% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{extend_vec_last_val}
\alias{extend_vec_last_val}
\title{extend_vec_last_val}
\usage{
extend_vec_last_val(vector, n)
}
\arguments{
\item{vector}{a vector to add n elements to, all of which have the same value as the last value in vector}

\item{n}{number of elements to concatenate}
}
\value{
a vector with length(vector) + n all the last n values have the same value as length(vector)
}
\description{
extend_vec_last_val
}
