% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{find_apportioned_fref}
\alias{find_apportioned_fref}
\title{find_apportioned_fref
deterministically search for an F that achieves some future Bzero value}
\usage{
find_apportioned_fref(proj_obj, proj_pars, percent_Bzero = 40)
}
\arguments{
\item{proj_obj}{an obj that has been created by \code{TMB::MakeADFun}. It is assumed this obj object is setup for projection e.g., \code{do_projection = 1} etc}

\item{proj_pars}{vector of input parameters that are passed to proj_obj$report()}

\item{percent_Bzero}{This is the target biomass that will trigger the Fref value reported.}
}
\value{
list with an F that has solved for a global F to achieve some percent_Bzero
}
\description{
find_apportioned_fref
deterministically search for an F that achieves some future Bzero value
}
\details{
This algorithm will search over a range of F values to find one that closely results in percent B0 specified by percent_Bzero in the terminal projection year.
Consider using \code{setup_proj_data} when taking an estimation model and setting up a \code{proj_obj}
}
