% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{find_fref}
\alias{find_fref}
\title{find_fref
deterministically search for an F that achieves some future Bzero value}
\usage{
find_fref(mle_obj, percent_Bzero = 40)
}
\arguments{
\item{mle_obj}{an optimised obj that has been created by \code{TMB::MakeADFun}}

\item{percent_Bzero}{This is the target biomass that will trigger the Fref value reported.}
}
\value{
list with an F that has solved for a global F to achieve some percent_Bzero
}
\description{
find_fref
deterministically search for an F that achieves some future Bzero value
}
\details{
This algorithm will search over a range of F values to find one that closely results in percent B0 specified by percent_Bzero in the terminal projection year. .It is best practice to include future uncertainty when searching for Fmsy, it comes with considerable computational cost.
}
