% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{find_regional_Fspr}
\alias{find_regional_Fspr}
\title{find_regional_Fspr
deterministically search for a vector of Fs (a value for each region) that achieves some regional percent b0}
\usage{
find_regional_Fspr(
  data,
  MLE_report,
  n_years_for_fleet_ratio = 5,
  percent_Bzero = 40,
  n_future_years = 60,
  verbose = FALSE
)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{n_years_for_fleet_ratio}{an integer specifying how many terminal years to average over to calculate F proportion by gear within each region}

\item{percent_Bzero}{This is the target biomass that will trigger the Fref value reported.}

\item{n_future_years}{an integer, specifying how long to simulate ahead by for calculations. Must be long enough for the model to reach som convergence.}

\item{verbose}{boolean print information as you are trialling Fs to help debug algorithm}
}
\value{
list with estimated Fs that have been solved to achieve some percent_Bzero in all regions
}
\description{
find_regional_Fspr
deterministically search for a vector of Fs (a value for each region) that achieves some regional percent b0
}
\details{
This algorithm will search over a vector of  F values for each region to find values that result in percent B0 specified by percent_Bzero in the terminal projection year.
Consider using \code{setup_proj_data} when taking an estimation model and setting up a \code{proj_obj}. The ratio of F among fishing fleets within a region is specified by the input parameter \code{n_years_for_fleet_ratio}
}
