% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{fix_pars}
\alias{fix_pars}
\title{fix_pars}
\usage{
fix_pars(
  par_list,
  pars_to_exclude,
  vec_elements_to_exclude = NULL,
  array_elements_to_exclude = NULL
)
}
\arguments{
\item{par_list}{a named list that you give to the par argument in the MakeADFun}

\item{pars_to_exclude}{a vector of strings with names of parameters you want to FIX in the objective object.}

\item{vec_elements_to_exclude}{a named list (names \%in\% pars_to_exclude) with number of elements = length(vec_pars_to_adjust). each list element}

\item{array_elements_to_exclude}{a named list (names \%in\% pars_to_exclude) with a matrix each row corresponds to an element with the first column being the array row index and second column being the array column index to fix
contains a vector of elements that we want to exclude from estimation.}
}
\value{
a list of factors used in the MakeADFun function
}
\description{
TMB helper function this function returns a list of factors used in the map argument of the MakeADFun function
values with NA will not be estimated.
}
\author{
C.Marsh
}
