% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{Francis_reweighting}
\alias{Francis_reweighting}
\title{Francis_reweighting}
\usage{
Francis_reweighting(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
named list length_multipliers
\itemize{
\item length_multipliers stage 2 weights for length observations
\item age_multipliers  stage 2 weights for age observations
\item mean_age_df mean age info used to calculate weights
\item mean_len_df mean length info used to calculate weights
}
}
\description{
Francis_reweighting
}
\details{
Calculates stage two weights for each region and observation using method TA1.8 from \insertCite{francis2011data}{SpatialSablefishAssessment}
}
\references{
\insertAllCited{}
}
