% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{get_comp_sample_size}
\alias{get_comp_sample_size}
\title{get_comp_sample_size}
\usage{
get_comp_sample_size(MLE_report, data, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{data}{list of input data}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data.frame
}
\description{
get_comp_sample_size
}
\details{
get compositional input sample sizes and effective sample sizes
}
