% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{get_fishing_mortalities}
\alias{get_fishing_mortalities}
\title{get_fishing_mortalities}
\usage{
get_fishing_mortalities(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data frame in long format
}
\description{
get_fishing_mortalities
}
