% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{get_init_nage}
\alias{get_init_nage}
\title{get_init_nage get initial numbers at age}
\usage{
get_init_nage(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{data}{list that is passed to the MakeADfun for the TMB model}
}
\value{
data.frame in long format
}
\description{
get_init_nage get initial numbers at age
}
