% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputFunctions.R
\name{get_input_observations}
\alias{get_input_observations}
\title{get_input_observations}
\usage{
get_input_observations(data, region_key = NULL)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data.frame when observation occurs in a year and region
}
\description{
get_input_observations
}
