% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInputValidation.R
\name{get_max_sel_pars}
\alias{get_max_sel_pars}
\title{get_max_sel_pars}
\usage{
get_max_sel_pars(sel_type_for_blocks)
}
\arguments{
\item{sel_type_for_blocks}{vector of selecvitiy types}
}
\value{
integer specifying the number of parameters for this selectivity parameter
}
\description{
get_max_sel_pars
}
