% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{get_mean_age}
\alias{get_mean_age}
\title{get_mean_age}
\usage{
get_mean_age(
  MLE_report,
  observation = "fixed",
  subset_years = NULL,
  sex = "both",
  region_key = NULL
)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{observation}{character labeling the observation you want to plot. See below for options
\itemize{
\item all
\item trwl
\item fixed
\item srv_dom_ll
}}

\item{subset_years}{vector of years to plot it for}

\item{sex}{character that allows users to specify if the want sex specific plots}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
a data frame of mean ages
}
\description{
get_mean_age
}
