% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{get_movement}
\alias{get_movement}
\title{get_movement}
\usage{
get_movement(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{obj$report()}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data frame
}
\description{
get_movement
}
