% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_model_run_comparions.R
\name{get_multiple_Finits}
\alias{get_multiple_Finits}
\title{get_multiple_Finits}
\usage{
get_multiple_Finits(mle_ls, run_labels = NULL)
}
\arguments{
\item{mle_ls}{list with multiple obj$report() calls}

\item{run_labels}{vector of strings that are labels for each element in mle_ls}
}
\value{
data frame with multiple F initial parameters
}
\description{
get_multiple_Finits
}
