% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_model_run_comparions.R
\name{get_multiple_input_datasets}
\alias{get_multiple_input_datasets}
\title{get_multiple_input_datasets accessors to get multiple input_datasets}
\usage{
get_multiple_input_datasets(data_ls, run_labels = NULL, region_key = NULL)
}
\arguments{
\item{data_ls}{list with multiple data lists}

\item{run_labels}{vector of strings that are labels for each element in mle_ls}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data frame with multiple negative log-likelihoods
}
\description{
get_multiple_input_datasets accessors to get multiple input_datasets
}
