% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{get_n_datasets}
\alias{get_n_datasets}
\title{get_n_datasets get a data frame of the number of data sets}
\usage{
get_n_datasets(MLE_report)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}
}
\value{
data frame with age-frequency info
}
\description{
get_n_datasets get a data frame of the number of data sets
}
