% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{get_negloglike}
\alias{get_negloglike}
\title{get_negloglike get a data frame of negative log likelihoods}
\usage{
get_negloglike(MLE_report)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}
}
\value{
data frame with age-frequency info
}
\description{
get_negloglike get a data frame of negative log likelihoods
}
