% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInputValidation.R
\name{get_number_of_sel_pars}
\alias{get_number_of_sel_pars}
\title{get_number_of_sel_pars}
\usage{
get_number_of_sel_pars(sel_type)
}
\arguments{
\item{sel_type}{integer of selecvitiy types}
}
\value{
integer specifying the number of parameters for this selectivity parameter
}
\description{
get_number_of_sel_pars
}
