% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{get_other_derived_quantities}
\alias{get_other_derived_quantities}
\title{get_other_derived_quantities}
\usage{
get_other_derived_quantities(MLE_report, data, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
A list of
}
\description{
get_other_derived_quantities
}
