% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{get_partition}
\alias{get_partition}
\title{get_partition}
\usage{
get_partition(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data frame of parition
}
\description{
get_partition
}
