% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{get_selectivities}
\alias{get_selectivities}
\title{get_selectivities}
\usage{
get_selectivities(MLE_report)
}
\arguments{
\item{MLE_report}{obj$report()}
}
\value{
data frame
}
\description{
get_selectivities
}
