% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{get_tag_recovery_obs_fitted_values}
\alias{get_tag_recovery_obs_fitted_values}
\title{get_tag_recovery_obs_fitted_values}
\usage{
get_tag_recovery_obs_fitted_values(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data frame in long format
}
\description{
get_tag_recovery_obs_fitted_values
}
