% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{get_tag_release_AF}
\alias{get_tag_release_AF}
\title{get_tag_release_AF}
\usage{
get_tag_release_AF(data, region_key = NULL)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
get_tag_release_AF
}
