% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{get_tag_release_ndx}
\alias{get_tag_release_ndx}
\title{get_tag_release_ndx  utility function to get release event ndx given release region and release year}
\usage{
get_tag_release_ndx(region_ndx, release_event_year_ndx, n_regions)
}
\arguments{
\item{region_ndx}{region index 1:n_regions}

\item{release_event_year_ndx}{release year index 1:n_years_to_retain_tagged_cohorts_for. 1 indicates tagged released this year, 2 is tags released last year etc.}

\item{n_regions}{number of regions in the model}
}
\value{
a tag release event index
}
\description{
get_tag_release_ndx  utility function to get release event ndx given release region and release year
}
