% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{get_tag_reporting_rate}
\alias{get_tag_reporting_rate}
\title{get_tag_reporting_rate get a data frame of reporting rate parameters}
\usage{
get_tag_reporting_rate(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
data.frame of tag-reporting rates by tag-recovery year recovery region
}
\description{
get_tag_reporting_rate get a data frame of reporting rate parameters
}
