% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{get_tmb_fixed_effects}
\alias{get_tmb_fixed_effects}
\title{get_tmb_fixed_effects}
\usage{
get_tmb_fixed_effects(obj)
}
\arguments{
\item{obj}{An optimised list that has been build by MakeAdFun}
}
\description{
return MLE estaimtes for fixed effects
}
