% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{get_tmb_parameter_element}
\alias{get_tmb_parameter_element}
\title{get_tmb_parameter_element utility function for turning an array index into a vector element.}
\usage{
get_tmb_parameter_element(parameters, element, parameter_label)
}
\arguments{
\item{parameters}{a named list that is passed to TMB::MakeADFun}

\item{element}{this is the element that is will be profiled in na_map$profile_param}

\item{parameter_label}{string that corresponds to a parameter label in parameters}
}
\value{
an integer that corresponds to how TMB vectorises array parameters
}
\description{
get_tmb_parameter_element utility function for turning an array index into a vector element.
}
\details{
TMB converts matrices and arrays into vectors internally, this function will take an vector or array index (element) and give TMBs vector location for that parameter
}
