% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{get_TMB_vector_from_array}
\alias{get_TMB_vector_from_array}
\title{get_TMB_vector_from_array}
\usage{
get_TMB_vector_from_array(element, array)
}
\arguments{
\item{element}{a vector of integers specifying the element of an array, length represents the array dimension}

\item{array}{the array that we are searching for}
}
\value{
element of vector
}
\description{
get_TMB_vector_from_array
}
