% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{gm_mean}
\alias{gm_mean}
\title{gm_mean calculate geometric mean when you have na's in vector}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of strictly positive values}

\item{na.rm}{boolean on whether to ignore NA's}
}
\value{
geometric mean with na's ignored
}
\description{
gm_mean calculate geometric mean when you have na's in vector
}
