% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{inv_logis_sel}
\alias{inv_logis_sel}
\title{inv_logis_sel Inverse Logisitic selectivity}
\usage{
inv_logis_sel(bins, a50, ato95)
}
\arguments{
\item{bins}{could be age or length}

\item{a50}{bin which equates to selectivy being at 0.5}

\item{ato95}{bins between the selectivty 0.5 and 0.95}
}
\value{
selectivity for each age.
}
\description{
inv_logis_sel Inverse Logisitic selectivity
}
