% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{invlogit}
\alias{invlogit}
\title{invlogit Inverse logit transformation, equivalent to plogis(Y)}
\usage{
invlogit(Y)
}
\arguments{
\item{Y}{scalar between -inf, inf}
}
\value{
X between \link{0,1}
}
\description{
invlogit Inverse logit transformation, equivalent to plogis(Y)
}
