% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{invlogit_general}
\alias{invlogit_general}
\title{invlogit_general bounds X which is between -inf -> inf to \link{lb,ub} based on the logit transformation}
\usage{
invlogit_general(Y, lb, ub)
}
\arguments{
\item{Y}{scalar range \link{-inf, inf}}

\item{lb}{lower bound for X}

\item{ub}{upper bound for X}
}
\value{
X to be between \link{lb,ub}
}
\description{
invlogit_general bounds X which is between -inf -> inf to \link{lb,ub} based on the logit transformation
}
