% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{is_constant}
\alias{is_constant}
\title{is_constant
is a vector constant}
\usage{
is_constant(x, tol = .Machine$double.eps)
}
\arguments{
\item{x}{vector}

\item{tol}{the magnitude that flags a difference}
}
\description{
is_constant
is a vector constant
}
