% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{is_matrix_invertable}
\alias{is_matrix_invertable}
\title{is_matrix_invertable}
\usage{
is_matrix_invertable(m)
}
\arguments{
\item{m}{an n x n matrix}
}
\value{
bool
\itemize{
\item false: not invertable
\item true: is invertable
}
}
\description{
helper function to see if a matrix is invertable
}
\examples{
\dontrun{
x <- matrix(rep(1,25),nc=5)          # singular
y <- matrix(1+1e-10*rnorm(25),nc=5)  # very nearly singular matrix
z <- 0.001*diag(1,5)                 # non-singular, but very smalll determinant
is_matrix_invertable(x)
# [1] FALSE
is_matrix_invertable(y)
# [1] TRUE
is_matrix_invertable(z)
# [1] TRUE
}
}
