% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{is_positive_definite}
\usage{
is_positive_definite(m, tol = 1e-06)
}
\arguments{
\item{m}{an n x n matrix}

\item{tol}{real value. Eigen values less than this tolerance value will fail the check.}
}
\value{
bool
\itemize{
\item false: not a positive definite matrix
\item true: is a positive definite matrix
}
}
\description{
helper function to see if a matrix is positive definite
}
