% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{knife_sel}
\alias{knife_sel}
\title{knife_sel Knife Edge selectivity}
\usage{
knife_sel(E, bins)
}
\arguments{
\item{E}{age at inflection}

\item{bins}{could be age or length}
}
\value{
selectivty for all ages
}
\description{
knife_sel Knife Edge selectivity
}
