% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{log_cv}
\alias{log_cv}
\title{log_cv Calculate the CV of the lognormal distribution based on \deqn{cv = \sqrt{e^{\sigma^2} - 1}}}
\usage{
log_cv(sigma)
}
\arguments{
\item{sigma}{The standard deviation of the lognormal distribution}
}
\value{
The the cv
}
\description{
log_cv Calculate the CV of the lognormal distribution based on \deqn{cv = \sqrt{e^{\sigma^2} - 1}}
}
