% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{log_sigma}
\alias{log_sigma}
\title{log_sigma}
\usage{
log_sigma(cv)
}
\arguments{
\item{cv}{The CV (note this is in proportion not percentage) of the lognormal distribution}
}
\value{
The the sigma
}
\description{
Calculate the sigma of the lognormal distribution based on \deqn{\sigma = \sqrt{log(cv^2 + 1)}}
}
