% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{logis}
\alias{logis}
\title{logis Logistic selectivity}
\usage{
logis(x, x50, a95, amax = 1)
}
\arguments{
\item{x}{age or length to evaluate selectivity at}

\item{x50}{value of x where selectivity is equal to 0.5}

\item{a95}{difference in x when the selectivity is equal to 0.5 and 0.95}
}
\value{
selectivity along x
}
\description{
logis Logistic selectivity
}
