% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selectivities.R
\name{logis_alt}
\alias{logis_alt}
\title{logis_alt an alternative logistic selectivity}
\usage{
logis_alt(x, x50, delta, amax = 1)
}
\arguments{
\item{x}{age or length to evaluate selectivity at}

\item{x50}{value of x where selectivity is equal to 0.5}

\item{delta}{delta parameter for the}
}
\value{
selectivity along x
}
\description{
logis_alt an alternative logistic selectivity
}
