% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{logit}
\alias{logit}
\title{logit bounds X which is between 0-1 to -inf -> inf based on the logit transformation
equivalent to qlogis(X)}
\usage{
logit(X)
}
\arguments{
\item{X}{scalar range \link{0,1}}
}
\value{
Y to be between -inf, inf
}
\description{
logit bounds X which is between 0-1 to -inf -> inf based on the logit transformation
equivalent to qlogis(X)
}
