% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{logit_general}
\alias{logit_general}
\title{logit_general bounds X which is between \link{lb,ub} to -inf -> inf based on the logit transformation}
\usage{
logit_general(X, lb, ub)
}
\arguments{
\item{X}{scalar range \link{lb,ub}}

\item{lb}{lower bound for X}

\item{ub}{upper bound for X}
}
\value{
Y to be between \link{-inf, inf}
}
\description{
logit_general bounds X which is between \link{lb,ub} to -inf -> inf based on the logit transformation
}
