% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{plot_AF}
\alias{plot_AF}
\title{plot_AF}
\usage{
plot_AF(
  MLE_report,
  observation = "srv_dom_ll",
  subset_years = NULL,
  sex = "both",
  region_key = NULL
)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{observation}{character labeling the observation you want to plot. See below for options
\itemize{
\item \code{trwl}
\item \code{fixed}
\item \code{srv_dom_ll}
}
If \code{MLE_report$model_type == 0} i.a., assessment model
\itemize{
\item \code{nmfs}
\item \code{srv_dom_ll}
\item \code{srv_jap_ll}
\item \code{fixed_gear}
}}

\item{subset_years}{vector of years to plot it for}

\item{sex}{character that allows users to specify if the want sex specific plots
\itemize{
\item \code{male}
\item \code{female}
\item \code{both}
}}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_AF
}
