% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputFunctions.R
\name{plot_age_error_matrix}
\alias{plot_age_error_matrix}
\title{plot_age_error_matrix}
\usage{
plot_age_error_matrix(data, fill_text = F)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{fill_text}{add value in ggplot}
}
\value{
ggplot2 object that will plot the ageing error matrix
}
\description{
plot_age_error_matrix
}
