% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputFunctions.R
\name{plot_age_length_matrix}
\alias{plot_age_length_matrix}
\title{plot_age_length_matrix}
\usage{
plot_age_length_matrix(data, subset_years = NULL)
}
\arguments{
\item{data}{list input for model}

\item{subset_years}{vector of years to subset the age-length plt for}
}
\value{
ggplot2
}
\description{
plot_age_length_matrix
}
