% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{plot_catch_fit}
\alias{plot_catch_fit}
\title{plot_catch_fit}
\usage{
plot_catch_fit(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_catch_fit
}
