% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{plot_comp_sample_size}
\alias{plot_comp_sample_size}
\title{plot_comp_sample_size}
\usage{
plot_comp_sample_size(MLE_report, data, region_key = NULL, plot_n_eff = T)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{data}{list of input data}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{plot_n_eff}{boolean if T plot effective sample size else plot input sample size}
}
\value{
data.frame
}
\description{
plot_comp_sample_size
}
\details{
plot compositional input sample sizes or effective sample sizes
}
