% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{plot_frequency_of_tag_release_and_recoveries}
\alias{plot_frequency_of_tag_release_and_recoveries}
\title{plot_tag_release_and_recoveries}
\usage{
plot_frequency_of_tag_release_and_recoveries(
  data,
  region_key = NULL,
  release_ndx_to_plot = 1:10,
  release_region_to_plt = NULL
)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{release_ndx_to_plot}{vector of integers to create subset plots}

\item{release_region_to_plt}{string matching release region to subset the plot for}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_tag_release_and_recoveries
}
