% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{plot_init_nage}
\alias{plot_init_nage}
\title{plot_init_nage plot initial numbers at age}
\usage{
plot_init_nage(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{data}{list that is passed to the MakeADfun for the TMB model}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_init_nage plot initial numbers at age
}
