% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputFunctions.R
\name{plot_input_observations}
\alias{plot_input_observations}
\title{plot_input_observations}
\usage{
plot_input_observations(data, region_key = NULL)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_input_observations
}
