% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{plot_mean_age}
\alias{plot_mean_age}
\title{plot_mean_age}
\usage{
plot_mean_age(
  MLE_report,
  observation = "fixed",
  subset_years = NULL,
  sex = "both",
  region_key = NULL,
  survey_labels = NULL
)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{observation}{character labeling the observation you want to plot. See below for options
\itemize{
\item all
\item trwl
\item fixed
\item srv_dom_ll
}}

\item{subset_years}{vector of years to plot it for}

\item{sex}{character that allows users to specify if the want sex specific plots
\itemize{
\item \code{male}
\item \code{female}
\item \code{both}
}}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{survey_labels}{character vector for each of the n_surveys}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_mean_age
}
