% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{plot_mean_length}
\alias{plot_mean_length}
\title{plot_mean_length}
\usage{
plot_mean_length(
  MLE_report,
  label = "fixed",
  subset_years = NULL,
  sex = "both",
  region_key = NULL
)
}
\arguments{
\item{label}{character labeling the observation you want to plot. See below for options
\itemize{
\item all
\item trwl
\item fixed
}}

\item{subset_years}{vector of years to plot it for}

\item{sex}{character that allows users to specify if the want sex specific plots}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{plot_mean_length}{a list that is output from obj$report() usually once an optimsation routine has been done.}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_mean_length
}
