% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputFunctions.R
\name{plot_mean_weight}
\alias{plot_mean_weight}
\title{plot_mean_weight}
\usage{
plot_mean_weight(data)
}
\arguments{
\item{data}{list input for model}
}
\value{
ggplot2
}
\description{
plot_mean_weight
}
