% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{plot_movement}
\alias{plot_movement}
\title{plot_movement}
\usage{
plot_movement(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{obj$report()}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2
}
\description{
plot_movement
}
