% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionFunctions.R
\name{plot_partition}
\alias{plot_partition}
\title{plot_partition plot numbers at age over time and space}
\usage{
plot_partition(MLE_report, subset_years = NULL, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{subset_years}{if you only want to plot the partition for some years}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_partition plot numbers at age over time and space
}
