% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{plot_selectivities}
\alias{plot_selectivities}
\title{plot_selectivities}
\usage{
plot_selectivities(MLE_report)
}
\arguments{
\item{MLE_report}{obj$report()}
}
\value{
ggplot2
}
\description{
plot_selectivities
}
