% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{plot_SSB}
\alias{plot_SSB}
\title{plot_SSB}
\usage{
plot_SSB(MLE_report, region_key = NULL, depletion = F)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{depletion}{boolean if true we will scale values by Bzero}

\item{data}{list that is passed to the MakeADfun for the TMB model}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_SSB
}
