% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{plot_tag_recovery_fits}
\alias{plot_tag_recovery_fits}
\title{plot_tag_recovery_fits plot predicted and observed tag-recovery obs}
\usage{
plot_tag_recovery_fits(MLE_report, region_key = NULL, plt_type = "aggregate")
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{plt_type}{string for the plot type
\itemize{
\item \code{aggregate}: aggregated observed and predicted
\item \code{stand_resid}: boxplots of standardised residuals for release region and recovery region i.e., boxplots are all the release-events
\item \code{rqr_resid}: boxplots of randomised quantile residuals for release region and recovery region i.e., boxplots are all the release-events
\item \code{raw_resid}: points and lines of raw residuals by recovery year (aggregated over release events) facets are release region by recovery region i.e., can by multiple release years for a recovery year
}}
}
\value{
ggplot2 facets are columns recovery regions rows are release regions
}
\description{
plot_tag_recovery_fits plot predicted and observed tag-recovery obs
}
