% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{plot_tag_release_AF}
\alias{plot_tag_release_AF}
\title{plot_tag_release_AF}
\usage{
plot_tag_release_AF(data, region_key = NULL, release_ndx_to_plot = 1:10)
}
\arguments{
\item{data}{list that is passed to the MakeADfun for the TMB model}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}

\item{release_ndx_to_plot}{vector of integers to create subset plots}
}
\value{
ggplot2 object that will plot if an observation occurs in a year and region
}
\description{
plot_tag_release_AF
}
