% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TagFunctions.R
\name{plot_tag_reporting_rate}
\alias{plot_tag_reporting_rate}
\title{plot_tag_reporting_rate create a ggplot of reporring rate by recovery year and region}
\usage{
plot_tag_reporting_rate(MLE_report, region_key = NULL)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
ggplot of reporting rates
}
\description{
plot_tag_reporting_rate create a ggplot of reporring rate by recovery year and region
}
