% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{posfun}
\alias{posfun}
\title{posfun
utility function to mimic TMB code and likelihood evaluations}
\usage{
posfun(x, eps, pen = NULL)
}
\arguments{
\item{x}{a value to check if its greater than epx}

\item{eps}{a small value to check if x is greater than}

\item{pen}{a value which will get incremented by the penalty of the posfun}
}
\value{
a modified value of x which is larger than eps in a differential manor
}
\description{
posfun
utility function to mimic TMB code and likelihood evaluations
}
