% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{post_optim_sanity_checks}
\alias{post_optim_sanity_checks}
\title{post_optim_sanity_checks run a few sanity checks before estimating a model}
\usage{
post_optim_sanity_checks(mle_obj, mle_pars, max_abs_gradient = 1e-05)
}
\arguments{
\item{mle_obj}{which is TMB object that contains fn(), gr(), and par functions. Should have been optimised}

\item{mle_pars}{vector of fixed effect parameters that can be passed to fn and gr obj calls}
}
\value{
bool true means model passed false means model failed. Will also print a message to screen on what failed.
}
\description{
This function will do some simple checks such check likelihoods are all finite, no non-zero gradients
}
\author{
C.Marsh
}
