% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{pow}
\alias{pow}
\title{pow
utility function to mimic TMB code and likelihood evaluations}
\usage{
pow(val, exponent)
}
\arguments{
\item{val}{base value}

\item{exponent}{exponent to take to the power}
}
\value{
val to the exponent
}
\description{
pow
utility function to mimic TMB code and likelihood evaluations
}
