% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{pre_optim_sanity_checks}
\alias{pre_optim_sanity_checks}
\title{pre_optim_sanity_checks run a few sanity checks before estimating a model}
\usage{
pre_optim_sanity_checks(obj)
}
\arguments{
\item{obj}{an estimated TMB object that contains fn(), gr(), and par functions}
}
\value{
bool true means model passed false means model failed. Will also print a message to screen on what failed.
}
\description{
This function will do some simple checks such check likelihoods are all finite, no non-zero gradients
}
\author{
C.Marsh
}
