% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{profile_param}
\alias{profile_param}
\title{profile_param}
\usage{
profile_param(
  parameters,
  mle_obj,
  na_map,
  profile_param_label,
  element = 1,
  profile_values,
  same_element = NULL,
  no_estimation = F,
  verbose = T
)
}
\arguments{
\item{parameters}{a named list that is passed to TMB::MakeADFun. This is used as the starting values for the profile runs}

\item{mle_obj}{which is TMB object that contains fn(), gr(), and par functions. Should have been optimised}

\item{na_map}{a named list that was used during optimisation}

\item{profile_param_label}{string that corresponds to a parameter in na_map and mle_obj$par}

\item{element}{this is the element that corresponds to the index of parameters$profile_param. a single value means its a vector, a matrix with one row indicates the parameter is an array/matrix (can only handle 2-3 dimensional arrays)}

\item{profile_values}{numeric vector of values that the model will profile the parameter at}

\item{same_element}{this is the element of other values of parameters$profile_param that need to be set the same as element i.e., they are estimated at the same value. a vector of values means its a vector, a matrix with multiple row indicates the elements that will be set the same as the profile parameter.}

\item{no_estimation}{bool only should be true for unit-testing purposes or debugging}

\item{verbose}{print information during the function call}
}
\value{
a named list with the following objects
\itemize{
\item \code{profile_mle} a list each element is an estimated report for each profile value
\item \code{parameters_ls}: a list each element is
\item \code{na_map}: single value for all years and regions
\item \code{profile_values} vector of values that was profiled
}
}
\description{
profile_param
}
\details{
this will run a profile for a parameter or related set of parameters
}
