% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{Q_sum_to_zero_QR}
\alias{Q_sum_to_zero_QR}
\title{Q_sum_to_zero_QR
calculate QR vector for sum to zero hard constraint}
\usage{
Q_sum_to_zero_QR(N)
}
\arguments{
\item{N}{integer number of elements in vector that we want to sum = 0}
}
\value{
a vector length N * 2 that will be used by sum_to_zero_QR
}
\description{
Q_sum_to_zero_QR
calculate QR vector for sum to zero hard constraint
}
