% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGroomingCode.R
\name{record_grooming_rule}
\alias{record_grooming_rule}
\title{record_grooming_rule}
\usage{
record_grooming_rule(
  df,
  index,
  catch.col,
  record = NULL,
  rule,
  year.col = NULL,
  attribute = "catch",
  keep_NA = T
)
}
\arguments{
\item{df}{the df we are manipulating}

\item{index}{a conditional index length(index) == nrow(df). TRUE means it stays, FALSE means it is being excluded from}

\item{catch.col}{column label for the catch variable}

\item{record}{a data.frame to append the summaries to.}

\item{rule}{the label for the grooming rule (string)}

\item{year.col}{column label for the year variable. Can be ommited, if supplied the Data.frame will break down catch and number of events removed from each grooming rule by year.}

\item{attribute}{if by year specify the attribute to record
\itemize{
\item catch absoulte catch
\item events number of events
\item relative_catch is the percent change in catch from applying the rule
\item relative_events is the percent change in number of events from applying the rule
\item cumulative_catch track the catch left in the data set after each grooming rule
\item cumulative_events track the events left in the data set after each grooming rule
}}

\item{keep_NA}{if index has NA's and this is true then they stay in the data}
}
\value{
the record data frame with new entries for the grooming rule
}
\description{
record_grooming_rule
}
\details{
a utility function to record the effect of grooming rules on data sets. It will note remove the index from the dataset. TODO: As well as recording the effect of a groomong rule on catch and number of records summarise spatail distribution.
}
\examples{
\dontrun{
## example of plotting this output
## plot records
melt_total = melt(total_grooming_record, id.vars = "rule")
melt_total$rule = factor(melt_total$rule, ordered = T, levels = total_grooming_record$rule)
ggplot(melt_total \%>\% filter(variable \%in\%  c("events", "catch")),
      aes(y = value /1000, x = rule, group = 1)) +
 geom_line(size = 2, linetype = "dotted") +
 geom_point(size = 4, aes(col = rule)) +
 theme(axis.text.x = element_blank()) +
 ylab("") +
 ylim(0,1600) +
 facet_wrap(~variable) +
 xlab("")
}
}
