% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{run_multi_F}
\alias{run_multi_F}
\title{run_multi_F
This function is run by the \code{find_multi_fref} method}
\usage{
run_multi_F(trial_Fs, input_pars, target_percent_Bzero, proj_obj)
}
\arguments{
\item{trial_Fs}{a vector of F values to be trialed should have length n_regions * 2. The first n_regions are for the fixed gear fishery and the second set are for the trawl}

\item{input_pars}{vector of input parameters that are passed to proj_obj$report()}

\item{target_percent_Bzero}{regional percent B0 that we are trying to solve for in all regions}

\item{proj_obj}{an obj that has been created by \code{TMB::MakeADFun}}
}
\value{
sum of squares difference between the Bzero resulting from the trial F's and the value we are trying to solve for target_percent_Bzero
}
\description{
run_multi_F
This function is run by the \code{find_multi_fref} method
}
