% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{run_this_F}
\alias{run_this_F}
\title{run_this_F
This function is run by the \code{find_apportioned_fref} method}
\usage{
run_this_F(
  trial_F,
  input_pars,
  target_percent_Bzero,
  proj_obj,
  fixed_F_allocation_by_region,
  trwl_F_allocation_by_region
)
}
\arguments{
\item{trial_F}{F value to F that is being trialed}

\item{input_pars}{vector of input parameters that are passed to proj_obj$report()}

\item{target_percent_Bzero}{global percent B0 that we are trying to solve for}

\item{proj_obj}{an obj that has been created by \code{TMB::MakeADFun}}

\item{fixed_F_allocation_by_region}{vector of proportions to allocated to each region for the fixed gear fishery}

\item{trwl_F_allocation_by_region}{vector of proportions to allocated to each region for the trawl gear fishery}
}
\value{
sum of squares difference between the the trial F's Bzero and the value we are trying to solve for target_percent_Bzero
}
\description{
run_this_F
This function is run by the \code{find_apportioned_fref} method
}
\details{
the \code{trial_F} is allocated among both \code{fixed_F_allocation_by_region} and \code{trwl_F_allocation_by_region} which must sum = 0
}
