% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{set_pars_to_be_the_same}
\alias{set_pars_to_be_the_same}
\title{set_pars_to_be_the_same}
\usage{
set_pars_to_be_the_same(par_list, map, base_parameters, copy_parameters)
}
\arguments{
\item{par_list}{a named list that you give to the par argument in the MakeADFun}

\item{map}{a list of factors that has been created by fix_pars(). parameters that you want fixed to other values should be set to NA in this object}

\item{base_parameters}{a named list (names) each element contains one index that will be used to set the value in copy_parameters}

\item{copy_parameters}{a named list (names) each element contains one index that will be set equal to the corresponding base_parameters}
}
\value{
a list of factors used in the MakeADFun function
}
\description{
TMB helper function this function returns a list of factors used in the map argument of the MakeADFun function
values with the same factor level will be estimated as the same value
}
\details{
TMB will estimate parameters based on the index specified in by the map argument in MakeADFun
so parameters with the same factor in map will be estimated as the same value.
NOTE: this only works for within the same parameter variable It doesn't work across parameters variables.
}
\author{
C.Marsh
}
