% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimationFunctions.R
\name{set_up_parameters}
\alias{set_up_parameters}
\title{set_up_parameters utility function to help 'turn off' parameters and share estimated parameters across elements}
\usage{
set_up_parameters(
  data,
  parameters,
  na_map = NULL,
  srv_sel_first_param_shared_by_sex = F,
  srv_sel_second_param_shared_by_sex = F,
  fixed_sel_first_shared_by_sex = F,
  fixed_sel_second_shared_by_sex = F,
  trwl_sel_first_shared_by_sex = F,
  trwl_sel_second_shared_by_sex = F,
  recruit_dev_years_not_to_estimate = NULL,
  srv_q_spatial = F,
  tag_reporting_rate = "constant",
  est_init_F = F,
  est_catch_sd = F,
  est_movement = T
)
}
\arguments{
\item{data}{a list of data inputs for the model}

\item{parameters}{a list of parameter values for the model}

\item{na_map}{an existing map that has already had parameters turned off, not well tested}

\item{srv_sel_first_param_shared_by_sex}{bool, whether the first survey selectivity parameter is shared among male and female for all time-blocks}

\item{srv_sel_second_param_shared_by_sex}{bool, whether the second survey selectivity parameter is shared among male and female for all time-blocks}

\item{fixed_sel_first_shared_by_sex}{bool, whether the first fixed gear selectivity parameter is shared among male and female for all time-blocks}

\item{fixed_sel_second_shared_by_sex}{bool, whether the second fixed gear selectivity parameter is shared among male and female for all time-blocks}

\item{trwl_sel_first_shared_by_sex}{bool, whether the first trawl gear selectivity parameter is shared among male and female for all time-blocks}

\item{trwl_sel_second_shared_by_sex}{bool, whether the second trawl gear selectivity parameter is shared among male and female for all time-blocks}

\item{recruit_dev_years_not_to_estimate}{vector of years that are not estimated.}

\item{srv_q_spatial}{whether regional Q's exist}

\item{tag_reporting_rate}{how to deal with tag-reporting rate will be ignored if there are no tag-recovery observations.
\itemize{
\item off: not estimated
\item ignore: ignore this parameter
\item constant: single value for all years and regions
\item time: estimates a tag-reporting rate coefficient for all recovery years which is common across all regions
\item space: TODO - not implemented estimates an regional tag-reporting rate which is common across all recovery years
\item spatio-temporal: TODO - not implemented - estimates annual and regional tag-reporting rates
}}

\item{est_init_F}{bool whether you want to estimate an initial F}

\item{est_catch_sd}{bool whether you want to estimate the catch sd parameter}

\item{est_movement}{bool whether you want to estimate movement parameters}
}
\value{
a named list that can be used by the \code{map} input for the \code{TMB::MakeADFun} function. NAs mean parameters are not estimated and elements with the same factor level mean they will be estimated with a common coefficient i.e., shared parameters
}
\description{
set_up_parameters utility function to help 'turn off' parameters and share estimated parameters across elements
}
