% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE_functions.R
\name{simulate_future_data}
\alias{simulate_future_data}
\title{simulate_future_data}
\usage{
simulate_future_data(
  data,
  parameters,
  n_future_years,
  future_ll_Fs,
  future_trwl_Fs,
  future_rec_devs = NULL
)
}
\arguments{
\item{data}{TMB data list which is up for current year}

\item{parameters}{TMB parameter which is up for current year}

\item{n_future_years}{number of years to run into the future}

\item{future_ll_Fs}{Annual fishing mortalities for the Longline fishery}

\item{future_trwl_Fs}{Annual fishing mortalities for the Trawl fishery}

\item{future_rec_devs}{future recruitment devs these are in logspace. If NULL we simulate from the prior}
}
\value{
list with the following elements sim_data, data, sim_parameters
}
\description{
simulate_future_data
}
\details{
Only used when \code{data$model == "Assessment"}
Any year based input quantity will be extended into the future using the last years values i.e., mean weight, M, age-length transition etc.
}
\examples{
\dontrun{
# this loads in mock data and parameters
load(system.file("testdata", "MockAssessmentModel.RData",package="SpatialSablefishAssessment"))
n_future_years = 3
future_ll_Fs = rlnorm(n_future_years, log(0.3, 0.1))
future_trwl_Fs = rlnorm(n_future_years, log(0.3, 0.1))
future_rec_devs = rnorm(n_future_years, -0.5 * data$sigma_R^2, data$sigma_R)
# run function get simulated data
new_sim_data = simulate_future_data(data, parameters, n_future_years, future_ll_Fs, future_trwl_Fs,future_rec_devs)
}
}
