% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{simulate_observations}
\alias{simulate_observations}
\title{simulate_observations}
\usage{
simulate_observations(obj, n_sims = 200, region_key = NULL)
}
\arguments{
\item{obj}{A TMB object that has been build using \code{TMB::MakeADFun}}

\item{n_sims}{an integer specifying how many simulated data sets you want}

\item{region_key}{data.frame with colnames area and TMB_ndx for providing real region names to objects}
}
\value{
named list containing simualted observations for all key obsevations
}
\description{
simulate_observations
}
\details{
Simulate observations conditional on MLE estimates
}
