% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialSablefishAssessment-package.R
\docType{package}
\name{SpatialSablefishAssessment-package}
\alias{SpatialSablefishAssessment}
\alias{SpatialSablefishAssessment-package}
\title{SpatialSablefishAssessment: Sablefish stock assessment package}
\description{
This package contains TMB models and utility functions used in spatial sablefish research, as well as a version that mimics the 2022 stock assessment model.
}
\author{
\strong{Maintainer}: Craig Marsh \email{Craig.Marsh10@gmail.com}

Authors:
\itemize{
  \item Daniel Goethel
}

}
\keyword{internal}
