% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{summarise_LF_quant_resids}
\alias{summarise_LF_quant_resids}
\title{summarise_LF_quant_resids}
\usage{
summarise_LF_quant_resids(LF_sim_resids, sex = "M", obs_label = "")
}
\arguments{
\item{LF_sim_resids}{an LF object created from \code{calculate_simulated_residuals}}

\item{sex}{character specifying which sex to plot
\itemize{
\item \code{M}: Male
\item \code{F}: Female
}}

\item{obs_label}{a character specifying the title possible observation label}
}
\value{
a joint plot
}
\description{
summarise_LF_quant_resids
}
