% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationFunctions.R
\name{summarise_tag_quant_resids}
\alias{summarise_tag_quant_resids}
\title{summarise_tag_quant_resids}
\usage{
summarise_tag_quant_resids(tag_sim_resids, recovery_year = T)
}
\arguments{
\item{tag_sim_resids}{an LF object created from \code{calculate_simulated_residuals}}

\item{recovery_year}{boolean whether to plot by recovery year = T or release year = F}
}
\value{
a joint plot
}
\description{
summarise_tag_quant_resids
}
