% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectionFunction.R
\name{trail_Fspr}
\alias{trail_Fspr}
\title{trail_Fspr
This function is run by the \code{find_regional_Fspr} method to solve for an Fspr}
\usage{
trail_Fspr(
  log_trial_Fs,
  fixed_F_proportion,
  n_future_years = 60,
  init_numbers_at_age,
  R0,
  natural_mortality,
  maturity_weight_at_age,
  trwl_sel,
  fixed_sel,
  movement_matrix,
  target_percent_Bzero = 40,
  prop_Z_in_ssb = 0,
  verbose = F,
  do_recruits_move
)
}
\arguments{
\item{log_trial_Fs}{a vector of log F values to be trialed should have length n_regions.}

\item{fixed_F_proportion}{for each region (trawl is is 1 - fixed_F_proportion)}

\item{n_future_years}{an integer, specifying how long to simulate ahead by for calculations. Must be long enough for the model to reach som convergence.}

\item{init_numbers_at_age}{matrix of initial numbers at age first value is assumed to be equal to 1. dim n_age x n_region}

\item{R0}{vector of R0 parameters for each region}

\item{natural_mortality}{natural mortality vector for ages}

\item{maturity_weight_at_age}{vector of maturity and weight at age}

\item{trwl_sel}{vector of selectivities by age for the trawl fishery}

\item{fixed_sel}{vector of selectivities by age for the fixed gear fishery}

\item{movement_matrix}{matrix of annual movement}

\item{target_percent_Bzero}{regional percent B0 that we are trying to solve for in all regions}

\item{prop_Z_in_ssb}{scalar specifying the proportion of Z to account for when calculating SSB during the year}

\item{verbose}{boolean print information as you are trialling Fs to help debug algorithm}

\item{do_recruits_move}{1 = yes 0 = no}
}
\value{
sum of squares difference between the Bzero resulting from the trial F's and the value we are trying to solve for target_percent_Bzero
}
\description{
trail_Fspr
This function is run by the \code{find_regional_Fspr} method to solve for an Fspr
}
