% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInputValidation.R
\name{validate_input_data_and_parameters}
\alias{validate_input_data_and_parameters}
\title{validate_input_data_and_parameters given a list of data and parameters will the TMB model crash?}
\usage{
validate_input_data_and_parameters(data, parameters)
}
\arguments{
\item{data}{a list of data inputs for the model}

\item{parameters}{a list of parameter values for the models}
}
\value{
string telling you if there are problems or if you are good to go
}
\description{
validate_input_data_and_parameters given a list of data and parameters will the TMB model crash?
}
