% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{vonbert}
\alias{vonbert}
\title{vonbert applies the Von Bertalanffy age-length relationship}
\usage{
vonbert(age, K, L_inf, t0)
}
\arguments{
\item{age}{take an age look in globe scope for the rest of parameters}

\item{K}{growth rate parameter}

\item{L_inf}{asympototic length}

\item{t0}{age for length = 0}
}
\value{
mean length at age
}
\description{
vonbert applies the Von Bertalanffy age-length relationship
}
