% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{zerofun}
\alias{zerofun}
\title{zerofun}
\usage{
zerofun(x, delta)
}
\arguments{
\item{x}{value to check if is zero}

\item{delta}{how small the value is before we alter it}
}
\value{
adjusted x if x < delta
}
\description{
zerofun
}
