% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{zerofun_v}
\alias{zerofun_v}
\title{zerofun_v vectorised version of zerofun}
\usage{
zerofun_v(x, delta)
}
\arguments{
\item{x}{value to check if is zero}

\item{delta}{how small the value is before we alter it}
}
\value{
adjusted x if x < delta
}
\description{
zerofun_v vectorised version of zerofun
}
