% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivedQuantityFunctions.R
\name{get_qs}
\alias{get_qs}
\title{get_qs}
\usage{
get_qs(MLE_report)
}
\arguments{
\item{MLE_report}{a list that is output from obj$report() usually once an optimsation routine has been done.}
}
\value{
A data frame of catchabilities
}
\description{
get_qs
}
