% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{inv_bound_unit}
\alias{inv_bound_unit}
\title{inv_bound_unit constrains Y from -inf -> inf to be between -1 -> 1}
\usage{
inv_bound_unit(X)
}
\arguments{
\item{X}{scalar range -1,1}
}
\value{
Y to be between -inf, inf
}
\description{
inv_bound_unit constrains Y from -inf -> inf to be between -1 -> 1
}
