% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialSablefishAssessment-package.R
\docType{package}
\name{SpatialSablefishAssessment-package}
\alias{SpatialSablefishAssessment}
\alias{SpatialSablefishAssessment-package}
\title{SpatialSablefishAssessment: Spatial sablefish stock assessment model}
\description{
This package contains a TMB model used in spatial sablefish research, with utility functions and unit-tests.
}
\author{
\strong{Maintainer}: Craig Marsh \email{Craig.Marsh10@gmail.com}

}
\keyword{internal}
