% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialSablefishAssessment-package.R
\docType{package}
\name{SpatialSablefishAssessment-package}
\alias{SpatialSablefishAssessment}
\alias{SpatialSablefishAssessment-package}
\title{SpatialSablefishAssessment: Sablefish stock assessment research package}
\description{
This package contains TMB models and utility functions used in spatial sablefish research, aswell as a version that mimics the 2021 stock assessment model.
}
\author{
\strong{Maintainer}: Craig Marsh \email{Craig.Marsh10@gmail.com}

}
\keyword{internal}
