% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BookdownFunctions.R
\name{summarise_multiple_models}
\alias{summarise_multiple_models}
\title{summarise_multiple_models
A bookdown function that will summaries a suite of models in individual Sections of a single Rmarkdown file}
\usage{
summarise_multiple_models(
  model_dir,
  bookdown_dir,
  bookdown_labels,
  model_description = ""
)
}
\arguments{
\item{model_dir}{a vector of absolute paths were we can find model output to summarise}

\item{bookdown_dir}{an absolute path where we will build the Bookdown}

\item{bookdown_labels}{a vector of model labels}

\item{model_description}{A string that describes all the models that are pointed at by model_dir}
}
\value{
Creates a suite of Rmarkdown files in bookdown_dir and compiles an R markdown file
}
\description{
summarise_multiple_models
A bookdown function that will summaries a suite of models in individual Sections of a single Rmarkdown file
}
